Better parse errors
Assume snippets of recipe texts are utf8 (they should be) to marginally
improve parser exceptions.

It'd be nice if memoryview slices would allow access to the offset of
the slice into the memoryview itself. Then it'd be reasonably straight
forward to render the code snippet where parsing is failing. I don't
think the memoryview API allows access to that. Maybe it'd be worth to
try write a custom implementation.