Make structure parsed into configurable
I'd like to be able to change the structure of the instruction text
generated, maybe even generate html for it eventually to allow things
like timers wrapped in an html component so they can be created with
behavior.

This commit allows a class to be passed in that will be used to build up
the recipe based on parsed elements.

It might be possible to create neater tests with this flexibility in
place. Something to think about.